Add Qt application and related QML files
Several QML files have been added to create a Qt application. The files include main.qml, which serves as the entry point of the application, along with model and delegate files to handle the app's data and functional logic. An AlarmDialog component is also introduced, and configuration for visual properties of the of certain UI components are also added through the qtquickcontrols2.conf file. Adjustments and additions in CMakeLists file are made to include the new QML files. Changes also include updates to gitignore file, some documentation, and other support files.